printer-agent/requirements.txt
Flask==3.0.3
python-escpos==3.1
qrcode==7.4.2
Pillow==10.4.0


